New helper CommandlineCaller
Wraps sarge and allows asynchronous execution of commands incl. logging
to provided logging methods for stdout, stderr and command calls.

Refactored PipCaller to utilized that.
